Issue #1308: share ranking validators
